requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3

# Audio Processing
whisper>=1.1.10
numpy>=1.24.0
torch>=2.1.0
torchaudio>=2.1.0
sounddevice>=0.4.6
librosa>=0.10.1

# Performance Monitoring
prometheus-client>=0.17.1
psutil>=5.9.6

# Development Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
black>=23.10.1
ruff>=0.1.4

# Testing
pytest-cov>=4.1.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
loguru>=0.7.0

# Web framework
websockets>=11.0.0

# Database
supabase>=1.0.0

# Utilities
asyncio>=3.4.3
webrtcvad>=2.0.10 